TFG/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anomalib==2.0.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
Django==5.2.1
django-allauth==65.8.1
docstring_parser==0.16
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
gradio==5.23.1
gradio_client==1.8.0
groovy==0.1.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jsonargparse==4.38.0
jwt==1.3.1
kiwisolver==1.4.8
lightning==2.5.1.post0
lightning-utilities==0.14.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
pycparser==2.22
pydantic==2.11.0
pydantic_core==2.33.0
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-argparse==1.7.0
ruff==0.11.2
safehttpx==0.1.6
semantic-version==2.10.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
svgwrite==1.4.3
sympy==1.13.1
tomlkit==0.13.2
torchmetrics==1.7.1
tqdm==4.67.1
Tree==0.2.4
typer==0.15.2
typeshed_client==2.7.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
yarl==1.20.0
psycopg==3.2.9
psycopg2==2.9.10
dj-database-url==2.3.0
